requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.10.0
future==1.0.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai-whisper==20240930
pydub==0.25.1
regex==2024.9.11
requests==2.32.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.6
triton==3.1.0
typing-extensions==4.12.2
urllib3==2.2.3
websockets==13.1
zipp==3.20.2
